SettingsResponse.revision changed from int into long

+++ android-sdk/src/main/java/com/sensorberg/sdk/internal/transport/model/SettingsResponse.java
@@ -8,7 +8,7 @@ public class SettingsResponse {
 
     @Expose
     @Getter
-    private int revision;
+    private long revision;
 
     @Expose
     @Getter
